# remove interior mutability of type-flags