--- a/src/librustc/traits/select.rs
+++ b/src/librustc/traits/select.rs
@@ -315,7 +315,7 @@ impl<'cx, 'gcx, 'tcx> SelectionContext<'cx, 'gcx, 'tcx> {
         self.infcx.tcx
     }
 
-    pub fn param_env(&self) -> &'cx ty::ParameterEnvironment<'gcx> {
+    pub fn param_env(&self) -> ty::ParameterEnvironment<'gcx> {
         self.infcx.param_env()
     }