… src/librustc/traits/README.md
… src/librustc/ty/mod.rs
… src/librustc/ty/util.rs
… src/librustc/util/ppaux.rs
… src/librustc_borrowck/borrowck/check_loans.rs
… src/librustc_borrowck/borrowck/mir/elaborate_drops.rs
… src/librustc_borrowck/borrowck/mir/gather_moves.rs
… src/librustc_borrowck/borrowck/mir/mod.rs
… src/librustc_borrowck/borrowck/mod.rs
… src/librustc_const_eval/check_match.rs
… src/librustc_lint/builtin.rs
… src/librustc_mir/build/mod.rs
… src/librustc_mir/hair/cx/mod.rs
… src/librustc_mir/shim.rs
… src/librustc_mir/transform/inline.rs
… src/librustc_mir/transform/qualify_consts.rs
… src/librustc_mir/transform/type_check.rs
… src/librustc_mir/util/elaborate_drops.rs
… src/librustc_passes/consts.rs
… src/librustc_trans/context.rs
… src/librustc_typeck/check/compare_method.rs
… src/librustc_typeck/check/dropck.rs
… src/librustc_typeck/check/method/probe.rs
… src/librustc_typeck/check/mod.rs
… src/librustc_typeck/check/regionck.rs
… src/librustc_typeck/coherence/builtin.rs
… src/test/run-make/atomic-lock-free/atomic_lock_free.rs