@nikomatsakis 2017-05-22
@bors r=eddyb